georoute/requirements.txt
# Core framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0

# HTTP clients
httpx>=0.26.0
requests>=2.31.0

# Google APIs
google-genai>=0.4.0
google-generativeai>=0.8.0

# Environment
python-dotenv>=1.0.0
pyyaml>=6.0.0

# Typing extensions (for Python 3.11 compatibility)
typing-extensions>=4.9.0

# Numpy FIRST - required before other scientific packages
numpy>=1.24.0,<2.0.0

# Terrain Data Processing (ESA WorldCover, Copernicus DEM)
rasterio>=1.3.0
scipy>=1.11.0
scikit-image>=0.21.0
shapely>=2.0.0
pyproj>=3.6.0
boto3>=1.34.0

# Building Detection & High-Resolution Pathfinding
# OSMnx for OpenStreetMap building footprints
osmnx>=1.9.0
geopandas>=0.14.0

# rdp for Douglas-Peucker path simplification
rdp>=0.8.0

# GPX export
gpxpy>=1.6.0

# SAM (Segment Anything Model) for building detection
# Note: torch and torchvision are pre-installed in the PyTorch Docker base image
segment-geospatial>=0.11.0

# OpenCV for image processing (SAM preprocessing)
opencv-python-headless>=4.8.0
Pillow>=10.0.0
